[EASY] Move Scorer to analysis tools (#1364)
* Move Scorer to analysis tools

* Fix import paths in tests and move test_scorer.py

* Run tox
