+++ snorkel/analysis/__init__.py
@@ -2,3 +2,4 @@
 
 from .error_analysis import get_label_buckets  # noqa: F401
 from .metrics import metric_score  # noqa: F401
+from .scorer import Scorer  # noqa: F401
